Revert "restructure dashboard-plot interaction"
This reverts commit 2b322404ec28ab3f84c857d57e0e89d2896523fc.
