+++ parmoo/viz/__init__.py
@@ -1,4 +1,2 @@
 from .plot import *
 from .dashboard import *
-from .graph import *
-from .settings import *
